bio/requirements.txt
biopython==1.84
boto3==1.34.131
ete3==3.1.3
jinja2==3.1.4
